pyproject.toml
[project]
name = "programming-rag"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "asyncpg>=0.31.0",
    "fastapi>=0.135.3",
    "langchain-community>=0.4.1",
    "langchain-core>=1.2.28",
    "langchain-google-genai>=4.2.1",
    "langchain-postgres>=0.0.17",
    "langgraph>=1.1.6",
    "langgraph-checkpoint-postgres>=3.0.5",
    "psycopg-pool>=3.3.0",
    "psycopg[binary]>=3.3.3",
    "pydantic>=2.12.5",
    "pypdf>=6.10.0",
    "python-dotenv>=1.2.2",
    "python-multipart>=0.0.26",
    "uvicorn[standard]>=0.44.0",
]
